chore: Move base.py outside from _estimator module
